+++ app/components/group_list/group_row_component.rb
@@ -1,12 +1,14 @@
 # frozen_string_literal: true
 
-module GroupListTree
+module GroupList
   # Represents a single row in the group list tree
   class GroupRowComponent < Component
     with_collection_parameter :group
+    attr_reader :group, :collapse
 
-    def initialize(group:)
+    def initialize(group:, collapse: true)
       @group = group
+      @collapse = collapse
     end
   end
 end
