--- app/components/group_list_tree_component.html.erb
@@ -1,1 +1,0 @@
-<div>Add GroupListTree template here</div>
