--- app/components/group_list_tree_component.rb
@@ -1,20 +1,0 @@
-# frozen_string_literal: true
-
-# This component renders a tree view of groups.
-class GroupListTreeComponent < ViewComponent::Base
-  erb_template <<-ERB
-      <div class="groups-list-tree-component">
-        <ul class="groups-list group-list-tree">
-          <%= render GroupListTree::GroupRowComponent.with_collection(@groups) %>
-        </ul>
-      </div>
-  ERB
-
-  def initialize(groups:)
-    @groups = groups
-  end
-
-  def render?
-    @groups.any?
-  end
-end
