+++ app/views/dashboard/groups/index.html.erb
@@ -12,15 +12,7 @@
 <div class="bg-white dark:bg-gray-800">
   <div class="flex flex-col">
     <div class="overflow-x-auto">
-      <%= render GroupListTreeComponent.new(groups: @groups) %>
-      <hr/>
-      <div class="shadow">
-        <div class="min-w-full">
-          <% @groups.each do |group| %>
-            <%= render group, depth: 0, collapse: false %>
-          <% end %>
-        </div>
-      </div>
+      <%= render GroupListComponent.new(groups: @groups) %>
     </div>
   </div>
 </div>
