+++ test/components/group_list_tree/group_component_test.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class GroupListTree::GroupComponentTest < ViewComponent::TestCase
+  def test_component_renders_something_useful
+    # assert_equal(
+    #   %(<span>Hello, components!</span>),
+    #   render_inline(GroupListTree::GroupComponent.new(message: "Hello, components!")).css("span").to_html
+    # )
+  end
+end
